requirements.txt
aniso8601==9.0.1
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.10
Werkzeug==1.0.1
argon2-cffi==23.1.0
python-dotenv==1.0.0
Flask-Mail==0.9.1
